managed user tokens for manga data,
for now we can have a list of client information (my data sources)

this was designed only to send emails/notifications towards leon.

if need be, we can change the way the set of clientIds, but considering
i am not planning to scale. we can just live with this and have a set of
clients and websites to detect for updates and cure data.